Created a new hook - usePartsOfPlanFunctions.ts to make PartsOfPlanTable.tsx component leaner.
